Fix bug of expectedEdges matrix only containing integer values

+++ graph.c
@@ -33,7 +33,7 @@ Graph *constructGraphFromInput(char *inputFilePath) {
 
     for (i = 0; i < n; i++) {
         for (j = 0; j < n; j++) {
-            setVal(G->expectedEdges, i, j, G->degrees[i] * G->degrees[j] / G->degreeSum);
+            setVal(G->expectedEdges, i, j, (double) G->degrees[i] * G->degrees[j] / G->degreeSum);
         }
     }
     fclose(graph_file);
